move component to shared components
