+++ src/components/v5/common/ActionSidebar/partials/Motions/steps/OutcomeStep/hooks.ts
@@ -1,7 +1,11 @@
 import { useMemo } from 'react';
+import { Id } from '@colony/colony-js';
 import { supportOption, opposeOption } from '../../consts';
 import { MotionVote } from '~utils/colonyMotions';
 import { ColonyMotion } from '~types';
+import { COLONY_TOTAL_BALANCE_DOMAIN_ID } from '~constants';
+import { useGetMembersForColonyQuery } from '~gql';
+import { useColonyContext } from '~hooks';
 
 export const useOutcomeStep = (motionData: ColonyMotion | null | undefined) => {
   const voteStatuses = useMemo(() => {
@@ -34,5 +38,29 @@ export const useOutcomeStep = (motionData: ColonyMotion | null | undefined) => {
     voteStatuses,
     yayPercent: motionData?.motionStakes?.percentage?.yay || '',
     nayPercent: motionData?.motionStakes?.percentage?.nay || '',
+  };
+};
+
+export const useMemberAvatars = (
+  currentDomainId = COLONY_TOTAL_BALANCE_DOMAIN_ID,
+) => {
+  const { colony } = useColonyContext();
+
+  const { data, loading } = useGetMembersForColonyQuery({
+    skip: !colony?.colonyAddress,
+    variables: {
+      input: {
+        colonyAddress: colony?.colonyAddress ?? '',
+        domainId: currentDomainId || Id.RootDomain,
+      },
+    },
+    fetchPolicy: 'cache-and-network',
+  });
+
+  const watchers = data?.getMembersForColony?.watchers;
+
+  return {
+    watchers: watchers || [],
+    loading,
   };
 };
