+++ src/components/v5/shared/MembersAvatars/types.ts
@@ -1,5 +1,9 @@
+import { Watcher } from '~types';
+
 export interface MembersAvatarsProps {
   currentDomainId?: number;
   maxAvatars?: number;
   className?: string;
+  loading: boolean;
+  watchers: Watcher[];
 }
